HIVE-24663 : Reduce overhead of partition column stats updation
