+++ standalone-metastore/metastore-server/src/test/java/org/apache/hadoop/hive/metastore/DummyRawStoreControlledCommit.java
@@ -1452,4 +1452,13 @@ public class DummyRawStoreControlledCommit implements RawStore, Configurable {
   public void dropPackage(DropPackageRequest request) {
     objectStore.dropPackage(request);
   }
+
+  @Override
+  public Map<String, Map<String, String>> updatePartitionColumnStatisticsInBatch(
+        Map<String, ColumnStatistics> partColStatsMap,
+        Table tbl, List<TransactionalMetaStoreEventListener> listeners,
+        String validWriteIds, long writeId)
+          throws MetaException, InvalidObjectException, NoSuchObjectException, InvalidInputException {
+    return objectStore.updatePartitionColumnStatisticsInBatch(partColStatsMap, tbl, listeners, validWriteIds, writeId);
+  }
 }
