+++ standalone-metastore/metastore-server/src/test/java/org/apache/hadoop/hive/metastore/DummyRawStoreForJdoConnection.java
@@ -1450,4 +1450,13 @@ public class DummyRawStoreForJdoConnection implements RawStore {
   @Override
   public void dropPackage(DropPackageRequest request) {
   }
+
+  @Override
+  public Map<String, Map<String, String>> updatePartitionColumnStatisticsInBatch(
+          Map<String, ColumnStatistics> partColStatsMap,
+          Table tbl, List<TransactionalMetaStoreEventListener> listeners,
+          String validWriteIds, long writeId)
+          throws MetaException, InvalidObjectException {
+    return null;
+  }
 }
